Creation of my own Repository for Submission
I cloned the repository for the project, but using gitHub desktop, did not seem to allow me to push my version up to website for submission. Not super happy about that as I made a fair amount of improvements and also changes to my style of submimitting notes. Lesson Learned though for next project. Include ReadMe, alternative controls and more.